add prettier

+++ src/BProtocolMorphoAllocator.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.2 <0.9.0;
 
-import { ISPythia, Signature, RiskData } from "./interfaces/ISPythia.sol";
-import { IMetaMorpho, MarketAllocation, Id, MarketParams } from "./interfaces/IMetaMorpho.sol";
-import { RiskyMath } from "./lib/RiskyMath.sol";
-import { MorphoLib } from "./lib/MorphoLib.sol";
+import {ISPythia, Signature, RiskData} from "./interfaces/ISPythia.sol";
+import {IMetaMorpho, MarketAllocation, Id, MarketParams} from "./interfaces/IMetaMorpho.sol";
+import {RiskyMath} from "./lib/RiskyMath.sol";
+import {MorphoLib} from "./lib/MorphoLib.sol";
 
 /*  
 USDC/sDAI
@@ -26,14 +26,14 @@ marketid: 0xbc6d1789e6ba66e5cd277af475c5ed77fcf8b084347809d9d92e400ebacbdd10
 */
 
 contract BProtocolMorphoAllocator {
-    using RiskyMath for uint;
+    using RiskyMath for uint256;
     using MorphoLib for MarketParams;
+
     ISPythia immutable SPYTHIA;
     address immutable TRUSTED_RELAYER;
     address immutable METAMORPHO_VAULT;
-    uint immutable MIN_CLF = 3;
+    uint256 immutable MIN_CLF = 3;
 
-    
     constructor(ISPythia spythia, address relayer, address morphoVault) {
         SPYTHIA = spythia;
         TRUSTED_RELAYER = relayer;
@@ -44,13 +44,17 @@ contract BProtocolMorphoAllocator {
         MarketAllocation[] memory allocations,
         RiskData[] memory riskDatas,
         Signature[] memory signatures
-    )
-        public
-    {
-        require(allocations.length == riskDatas.length, "Invalid number of risk data");
-        require(riskDatas.length == signatures.length, "Invalid number of signatures");
+    ) public {
+        require(
+            allocations.length == riskDatas.length,
+            "Invalid number of risk data"
+        );
+        require(
+            riskDatas.length == signatures.length,
+            "Invalid number of signatures"
+        );
 
-        for (uint i = 0; i < allocations.length; i++) {
+        for (uint256 i = 0; i < allocations.length; i++) {
             _checkAllocationRisk(allocations[i], riskDatas[i], signatures[i]);
         }
 
@@ -59,37 +63,50 @@ contract BProtocolMorphoAllocator {
     }
 
     function _checkAllocationRisk(
-    MarketAllocation memory allocation,
-    RiskData memory riskData,
-    Signature memory signature
+        MarketAllocation memory allocation,
+        RiskData memory riskData,
+        Signature memory signature
     ) private {
-        require(allocation.marketParams.collateralToken == riskData.collateralAsset, "Allocation collateral token != riskData.collateralAsset");
-        require(allocation.marketParams.loanToken == riskData.debtAsset, "allocation.loanToken != riskData.debtAsset");
-        
+        require(
+            allocation.marketParams.collateralToken == riskData.collateralAsset,
+            "Allocation collateral token != riskData.collateralAsset"
+        );
+        require(
+            allocation.marketParams.loanToken == riskData.debtAsset,
+            "allocation.loanToken != riskData.debtAsset"
+        );
+
         // Verify if the signature comes from the trusted relayer
-        address signer = SPYTHIA.getSigner(riskData, signature.v, signature.r, signature.s);
+        address signer = SPYTHIA.getSigner(
+            riskData,
+            signature.v,
+            signature.r,
+            signature.s
+        );
         require(signer == TRUSTED_RELAYER, "invalid signer");
 
         // get market config from the vault
         Id marketId = allocation.marketParams.id();
-        (uint184 cap , ,) = IMetaMorpho(METAMORPHO_VAULT).config(marketId);
+        (uint184 cap, , ) = IMetaMorpho(METAMORPHO_VAULT).config(marketId);
 
-
-        uint sigma = riskData.volatility;
-        uint l = riskData.liquidity;
-        uint d = cap; // supplyCap
-        uint beta = 15487; // liquidation bonus
+        uint256 sigma = riskData.volatility;
+        uint256 l = riskData.liquidity;
+        uint256 d = cap; // supplyCap
+        uint256 beta = 15487; // liquidation bonus
 
         // LTV  = e ^ (-c * sigma / sqrt(l/d)) - beta
-        uint cTimesSigma = MIN_CLF * sigma / 1e18;
-        uint sqrtValue = (1e18 * l / d).sqrt() * 1e9;
-        uint mantissa = (1 << 59) * cTimesSigma / sqrtValue;
+        uint256 cTimesSigma = (MIN_CLF * sigma) / 1e18;
+        uint256 sqrtValue = ((1e18 * l) / d).sqrt() * 1e9;
+        uint256 mantissa = ((1 << 59) * cTimesSigma) / sqrtValue;
 
-        uint expResult = mantissa.generalExp(59);
+        uint256 expResult = mantissa.generalExp(59);
 
-        uint recommendedLtv = (1e18 * (1 << 59)) / expResult - beta;
+        uint256 recommendedLtv = (1e18 * (1 << 59)) / expResult - beta;
 
         // check if the current ltv is lower or equal to the recommended ltv
-        require(recommendedLtv >= allocation.marketParams.lltv, "recommended ltv is lower than current lltv");
+        require(
+            recommendedLtv >= allocation.marketParams.lltv,
+            "recommended ltv is lower than current lltv"
+        );
     }
 }
